Events and scene manager
Added SceneManager_ (appended _ to avoid conflict with Unity SceneManager). Allows for multiple scenes to be loaded and cycled through at play time for debugging purposes.

Added framework for events.
GameEvent is a scriptable object that can be created from the right click menu. Each unique GameEvent can be added to a GameEventListener, and whenever the GameEvent is raised all listeners will respond.
GameEventListener is a MonoBehaviour that can be added to a GameObject as a component and is a listener for a specific event type. Currently there is support for 1, 2 and 3 parameter event listeners. Each listener can register a method to call when it receives notification of the event type that it listens for.

Added GlobalCameraOverride prefab. This is a temporary work around until SceneManager_ is updated further. When attached to a GameObject it will disabled the camera in the Global scene so that the console won't be spammed with a million warnings about there being multiple audio listeners.

Commit includes a number of other scripts and scenes created for testing purposes that aren't worth discussing at this time.