Update event system
Updated the event system with support for strongly typed listeners. This can be done by creating a new event listener and implementing any of the INParamEventListener interfaces, where N is some number of parameters. GameEvent will ensure that events are only dispatched to listeners that implement an interface that matches the parameters passed to the appropriate Raise method. The system can easily be extended to support listeners with more parameters as needed.