Created level reset functionality
A level restart function was needed. We tried using the existing SetActiveScene() method but this bugged out due to the way the level loading/unloading async operations were ordered. They were fine for changing to different scenes, but for reloading the same scene the SceneController was essentially immediately loading the existing copy of the same scene, and then unloading it. The level loading gods forbid trying to unload the world out from under your feet.

Refactored the SceneController. On a scene change it now fully unloads the current scene before loading the next. This may increase loading times slightly, but they were never that long anyway.