Finish first pass on MissionTimer and game saves
MissionTimer is a simple wrapper for a C# stopwatch and can be used to record the time a player takes to complete a mission. And instance of a mission timer is owned by GameManager and is publicly accessible.

A framework for game saves has been implemented. GameSave is a class that can be serialised/deserialised with json.net and can be updated as the projected progresses with the data that we need to save. For now it has some placeholder properties.

GameSaveUtility is a class for reading and writing GameSave objects. Saves are just json files so they can easily be edited by a player, but this is a low priority issue right now. GameManager has an instance of GameSaveUtility and will attempt to load a test game save when it starts.

Saves are stored at the path defined by Unity's Application.persistentDataPath. See here to see where that path maps to on different operating systems: https://docs.unity3d.com/ScriptReference/Application-persistentDataPath.html