First pass on civilian AI
Completed first pass on civilian AI.

Civilians can be one of two types, which is set on their state machine. They are:
* Idle - The citizen plays an idle animation forever
* Waypoints and Idle - Civilian will navigate waypoints set on their state machine. Civilian will optionally suspend their path to complete and idle animation if any are set in their waypoint actions.

Waypoint actions define an animation state to transition to, along with a timer for how long to stay in the state, and a probability that they will enter the state, which can introduce some randomness.

AI civilians are currently passive and have no way to detect the player. However, they use a modified version of the guard AI state machine and as such have all the references needed to add a FOV and a ThreatMeter to allow them sense and react to the player.

Animation states are defined by strings and must match the state name in the AI_Civilian animation controller. All states should have a matching enter and exit trigger. The current setup is such that all AI civilians can use the one AI_Civilian animation controller which allows for any number of idles animation states, but only one walking animation state. This can potentially be changed in the future.

Some updates have been made to the existing AI code to allow for the new civilian behaviours to work. Some of the changes are just simple null reference fixes on calls to a state machine context FOV as civilians do not currently have a FOV.

The biggest change is to the waypoint state. It now calls 'OnWaypointReached()' every time a waypoint is reached which allows for an AI State Machine context to react to a waypoint being reached. Whenever the waypoint state exits the AI navmesh agent is stopped, this prevents civilians from walking to their next destination, which will have been set but should not yet be navigated to if the civilian is transitioning to an idle animation playing state. Because of this change the other existing states have a call to start the navmesh agent so as to avoid any awkward stoppages on a state transition.

The civilian ai prefab has been added to Prefabs/Civilians. Dragging the prefab in should "just work". However, if you wish to change the model, you will need to unpack the prefab in the hierarchy, delete the model prefab, and then drag in the new model prefab. You will then need to change the animator's avatar reference to target the correct avatar for the civilian model.

The "Civilian Test" scene has an example of the default AI civilian prefab, named "AI_Civilian_Idle". There is also an unpacked prefab example "AI_Civilian_Waypoints" which has had the civilian model changed.