AmbientSound3D prefab first pass
First pass on AmvientSound3D prefab that can be placed throughout the game world.

Some example usages are in the WifeTest scene.

The prefab has an AudioSource and an AmbientSound3D script as components. The script can be used to setup the audio emitter with either a single looping AudioClip, or 1 or more AudioClips to play either in sequential or random order, and with an adjustable ranged gap between clips to introduce some randomness.

The only parameters necessary to adjust on the AudioSource are the base level volume of the source and the volume rolloff. Other settings on the source are available and can be adjusted as required.