Decrease sprint stamina decrease rate
Recent changes to use gravity on the player to prevent prop flying has mean that the player is now slightly slower than they used to be. The net effect was this it was impossible to sprint the distance to the final gate without the stamina depleting and the guard catching you unless your timing was frame perfect. Decreased the stamina decrease rate from 0.15 to 0.115 and the balance between tension and success feels much more suitable.