Adjust gate trigger position
Moved the gate close trigger position so that the player enters the trigger much sooner after clearing the gates after the stealable object room. This prevents the player from stopping short of the trigger and being caught by the guards, even when they are in a position where they should no longer get caught.

Fixes issue #69