Update checkpointing and fix some bugs
Checkpoint code now working as you would expect it to.
Pause menu and mission summary menu text will now read "restart level from checkpoint" when a checkpoint has been reached. When the player reaches the end of the level this will revert to "restart level" to indicate that the level will start over entirely.

Fixed missing collision mesh on ladder in library,

Rotated "Extraction Point" text 180 degrees so that it faces the direction the player would naturally read it from.

Adjusted the pope cart boxes slightly, but they still fall over when the level starts.

Fixed the trigger at the exit of the stealable object room not disabling itself.