Add "No Friction" material to critical gameplay sections
The flow of the game was getting bogged down in certain areas (most notable the sprint from the library to the party room door) due to the friction between the players collision capsule and the walls and bushes. Adding the no friction material to these walls means that the player can now glide along them while sprinting, rather than getting slowed down. This should help aid the player in not getting caught by guards as often in tight areas and sprint zones.